refactoring code
change components dir
