+++ app/(home)/[lang]/about-page.tsx
