+++ app/(home)/[lang]/activities-page.tsx
