+++ app/(home)/[lang]/parts-page.tsx
